pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personalab"
version = "0.1.0"
description = "Content factory library powered by Gemini: characters, images, videos, and content ideas."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = []
keywords = ["gemini", "content", "personalab", "llm", "image-generation", "video-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "google-genai>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "numpy>=1.24",
    "opencv-python>=4.8",
]

[project.optional-dependencies]
eval = [
    "insightface>=0.7",
    "onnxruntime>=1.16",
]
openai = ["openai>=1.0"]
runware = ["runware>=0.5", "httpx>=0.27"]
replicate = ["replicate>=0.25", "httpx>=0.27"]
runway = ["runwayml>=0.5", "httpx>=0.27"]
all-providers = [
    "personalab[openai,runware,replicate,runway]",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
personalab = ["data/**/*.yml", "data/**/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
